Màj et corrections Classe Hour: Màj des méthodes toString() et toQString() Classe LDCP et LDCR: Màj de l'opérateur [] Classe Manager: Màj des méthodes load et save Classe Personne: Correction des méthodes addRDV() et removeRDV() Classe RDV: Correction des méthodes addPersonne et removePersonne()

+++ Carnet_RDV/ldcp.h
@@ -17,7 +17,7 @@ private:
 class LDCP
 {
 private:
-    const unsigned PRINT_MODULO = 3;
+    const unsigned short PRINT_MODULO = 3;
 
     ChainonPersonne* d_t;
     unsigned Size;
